+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- Fixed race condition during tests parallel execution (`tests::ingest::test_fetch::*` test group)
+
 ## [0.150.0] - 2023-12-27
 ### Changed
 - Changed logic in `SimpleTransferProtocol` now block data and checkpoint downloading/uploading